Don't use access() to check for writability
The gnome-ostree build system uses a FUSE mount to access disk images
as non-root.  The files appear to be owned by root, but are actually
modifable by the build user.  Unfortunately, the access() system call
is unaware of this, and will error out.

An earlier patch changed the xml save error to write out a "please
run me as root" notification, so we'll still have the hint.

https://bugs.freedesktop.org/show_bug.cgi?id=61472