Fix misdetection of zip files as their content
Bump application/zip priority so that if the file is a PDF inside a zip
file, we don't have a toss up between those 2 to know which mime-type it
actually is.
Prefer the application/zip rather than whatever is inside.

But that means that all the priorities for the application/zip sub-types
magic need to be bumped to be higher than application/zip, so they are
detected as that type rather than zip.

In short: application/zip magic priority raised to 60
application/zip sub-types magic priorities raised to 70

This is similar to the fix in 10fc17e

https://bugs.freedesktop.org/show_bug.cgi?id=96660