Add text/x-python3 content type
This derives from the text/x-python content type, and is attached to
files with a .py3 extension, or which have a python3 shebang line.
Otherwise, text/x-python is preferred.

This is necessary because Python 3 is not compatible with Python 2, and
introduces some new keywords (such as ‘nonlocal’) which need to be
syntax highlighted differently (for example).

https://bugs.freedesktop.org/show_bug.cgi?id=100034