fix the QML magic
We can only say that some module starting with Qt or Qml will be imported
(as long as it's not QML for BB Cascades or some web framework or...)
and something will be declared with braces after.  The declaration must
start with a capital letter but we have no way to express that.
Alternatively, any script starting with shebang qml must be qml.

Added more tests.  The PySide minimal example is public domain, from
https://bitbucket.org/techtonik/pyside-copy-paste/src/5f8c1f8cf12d8fb18a1d35d98acc8200c904c56d/01.minimal.py?at=default&fileviewer=file-view-default

https://bugs.freedesktop.org/show_bug.cgi?id=104480