Raise SVG match priority to be above XML's
We lowered SVG's magic priority so that files would be first and
foremost detected as HTML, then as SVG. But SVG files are also XML, and
XML files had the same priority as SVGs, leading to SVG files being
detected as simply "XML".

Raise SVG's magic priority to be above XML, but below HTML's.
Fortunately we have a test now...

Fixes: 84563c62ab30190f8d3d25f8763556068f02d4a8