Adapt to xdgmime behaviour change.
As the spec says, if the result from magic sniffing doesn't
match one of the globs, ignore it and return one of the globs.

With this change, test-garbage.key (which was already a doubtful
testcase, see !5) no longer falls back to application/octet-stream
but to one of the globs.
test-really-html-not-xhtml.html is also fixed by this,
it's detected as text/html.

Closes: #138.