text/vnd.trolltech.linguist: require space or start tag closing > in magic
https://doc.qt.io/qt-5/linguist-ts-file-format.html

Avoids some more MPEG TS confusion. Hopefully that is -- I haven't
been able to locate MPEG TS samples that would be misdetected as
Linguist files in the first place. But because of that, I think it's
fair to assume that the sequence `<TS` is not that common in them, and
thus it followed by a space or the start tag closing `>` could
be even less common.

Not checking for `<TS/` as the root element may not be empty per the
DTD. Per [0] checking other whitespace besides the common space isn't
necessary either.

Refs https://gitlab.freedesktop.org/xdg/shared-mime-info/-/issues/10

[0] https://gitlab.freedesktop.org/xdg/shared-mime-info/-/merge_requests/103#note_741219