xdgmime: configure as submodule
So that we can bind it to a specific revision, and help with keeping
clones up to date.

GitLab CI instructs to use a relative path to submodules on the same
GitLab server [0], but doing so causes a meson/ninja dist failure [1].
Configure using a full https URL for now, as it appears to work with
GitLab just fine.

[0] https://docs.gitlab.com/ce/ci/git_submodules.html
[1] https://github.com/mesonbuild/meson/issues/8144