audio/vnd.dts.hd: include parent type magic in magic match
Matching just 0x64582025 within the first 18725 bytes of any file is
prone to false positives.

The priority of the magic match here being bumped to 60 seems to
indicate intent to make it take precedence over the parent
`audio/vnd.dts`. But we should take the parent magic into account here
as it does not "inherit" with `sub-class-of`.

As the comment for this match says "16bits HD be", accept only "16bits
core be" as the parent at least for now, not all possible parent
magics -- it would not seem consistent to have it in "14bits"
variants, or the "16bits core le" one. The test case we have for this
is one matching "16bits core be".

Refs https://gitlab.freedesktop.org/xdg/shared-mime-info/-/issues/166
Refs https://gitlab.freedesktop.org/xdg/shared-mime-info/-/issues/167