Move '*.blend' to top of Blender glob list
The globs for Blender scene files listed three possible matches,
with "*.blender" first, followed by two variations on "*.blend"
(lowercase and all-caps).

But Blender names its scene files "filename.blend", and always has.

Having "*.blender" listed first is problematic, because some tools
(e.g. GNOME Control Center) show only the first glob when describing
a type. As a result, its interface was claiming that Blender is the
default application for "*.blender" files, when it should be shown
as the default for "*.blend" files.